# api/docs/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MetivtaEval API",
    "description": "## MetivtaEval - AI Benchmarking Platform\n\nMetivtaEval provides evaluation services for Torah AI systems across DAAT and MTEB modes.\n\n### Features\n\n- **DAAT Scoring**: Deterministic Attribution & Agentic Traceability\n- **MTEB Metrics**: Standard information retrieval metrics\n- **Leaderboard APIs**: Ranked public and user-specific evaluation views\n- **WebSocket Events**: Authenticated progress and leaderboard update streams\n\n### Runtime Notes\n\n- **DAAT Dataset**: Loaded from the configured local JSON dataset in Docker and local dev\n- **LangSmith**: Optional for dataset sync and tracing when credentials are configured\n- **External Evaluators**: Anthropic and Browserless integrations activate when their keys are set\n\n### Authentication\n\nAll protected endpoints accept:\n- **Bearer Token**: JWT in the `Authorization` header\n- **API Key**: `X-API-Key` header",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/leaderboard/": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Get leaderboard",
        "description": "Return ranked leaderboard rows.",
        "operationId": "get_leaderboard_api_v2_leaderboard__get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LeaderboardMode",
              "default": "all"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/leaderboard/my-rankings": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Get my rankings",
        "description": "Return leaderboard entries owned by current user.",
        "operationId": "get_my_rankings_api_v2_leaderboard_my_rankings_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  },
                  "title": "Response Get My Rankings Api V2 Leaderboard My Rankings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/leaderboard/{system_id}": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Get system ranking",
        "description": "Return one leaderboard entry by system/evaluation id.",
        "operationId": "get_system_ranking_api_v2_leaderboard__system_id__get",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaderboardEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "system_id": {
            "type": "string",
            "format": "uuid",
            "title": "System Id"
          },
          "system_name": {
            "type": "string",
            "title": "System Name"
          },
          "system_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Version"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "overall_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Score"
          },
          "daat_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daat Score"
          },
          "ndcg_10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndcg 10"
          },
          "map_100": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map 100"
          },
          "mrr_10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mrr 10"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "system_id",
          "system_name",
          "system_version",
          "author",
          "organization",
          "mode",
          "overall_score",
          "daat_score",
          "ndcg_10",
          "map_100",
          "mrr_10",
          "dataset_name",
          "submitted_at"
        ],
        "title": "LeaderboardEntry",
        "description": "Single leaderboard row."
      },
      "LeaderboardMode": {
        "type": "string",
        "enum": [
          "all",
          "daat",
          "mteb"
        ],
        "title": "LeaderboardMode",
        "description": "Leaderboard mode filter."
      },
      "LeaderboardResponse": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/LeaderboardMode"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "entries",
          "total",
          "page",
          "page_size",
          "last_updated"
        ],
        "title": "LeaderboardResponse",
        "description": "Leaderboard list payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}